Creates package core-notifications-browser
