+++ package.json
@@ -223,6 +223,7 @@
     "@kbn/core-node-server": "link:bazel-bin/packages/core/node/core-node-server",
     "@kbn/core-node-server-internal": "link:bazel-bin/packages/core/node/core-node-server-internal",
     "@kbn/core-node-server-mocks": "link:bazel-bin/packages/core/node/core-node-server-mocks",
+    "@kbn/core-notifications-browser": "link:bazel-bin/packages/core/notifications/core-notifications-browser",
     "@kbn/core-overlays-browser": "link:bazel-bin/packages/core/overlays/core-overlays-browser",
     "@kbn/core-overlays-browser-internal": "link:bazel-bin/packages/core/overlays/core-overlays-browser-internal",
     "@kbn/core-overlays-browser-mocks": "link:bazel-bin/packages/core/overlays/core-overlays-browser-mocks",
@@ -850,6 +851,7 @@
     "@types/kbn__core-node-server": "link:bazel-bin/packages/core/node/core-node-server/npm_module_types",
     "@types/kbn__core-node-server-internal": "link:bazel-bin/packages/core/node/core-node-server-internal/npm_module_types",
     "@types/kbn__core-node-server-mocks": "link:bazel-bin/packages/core/node/core-node-server-mocks/npm_module_types",
+    "@types/kbn__core-notifications-browser": "link:bazel-bin/packages/core/notifications/core-notifications-browser/npm_module_types",
     "@types/kbn__core-overlays-browser": "link:bazel-bin/packages/core/overlays/core-overlays-browser/npm_module_types",
     "@types/kbn__core-overlays-browser-internal": "link:bazel-bin/packages/core/overlays/core-overlays-browser-internal/npm_module_types",
     "@types/kbn__core-overlays-browser-mocks": "link:bazel-bin/packages/core/overlays/core-overlays-browser-mocks/npm_module_types",
