+++ packages/BUILD.bazel
@@ -91,6 +91,7 @@ filegroup(
     "//packages/core/node/core-node-server-internal:build",
     "//packages/core/node/core-node-server-mocks:build",
     "//packages/core/node/core-node-server:build",
+    "//packages/core/notifications/core-notifications-browser:build",
     "//packages/core/overlays/core-overlays-browser-internal:build",
     "//packages/core/overlays/core-overlays-browser-mocks:build",
     "//packages/core/overlays/core-overlays-browser:build",
@@ -343,6 +344,7 @@ filegroup(
     "//packages/core/node/core-node-server-internal:build_types",
     "//packages/core/node/core-node-server-mocks:build_types",
     "//packages/core/node/core-node-server:build_types",
+    "//packages/core/notifications/core-notifications-browser:build_types",
     "//packages/core/overlays/core-overlays-browser-internal:build_types",
     "//packages/core/overlays/core-overlays-browser-mocks:build_types",
     "//packages/core/overlays/core-overlays-browser:build_types",
