+++ packages/core/notifications/core-notifications-browser/README.md
@@ -1,0 +1,3 @@
+# @kbn/core-notifications-browser
+
+This package contains the public types for core's browser-side notifications service.
