+++ packages/core/notifications/core-notifications-browser/jest.config.js
@@ -5,15 +5,9 @@
  * in compliance with, at your election, the Elastic License 2.0 or the Server
  * Side Public License, v 1.
  */
-export { NotificationsService } from './notifications_service';
-export type {
-  ErrorToastOptions,
-  ToastOptions,
-  Toast,
-  ToastInput,
-  IToasts,
-  ToastsApi,
-  ToastInputFields,
-  ToastsSetup,
-  ToastsStart,
-} from './toasts';
+
+module.exports = {
+  preset: '@kbn/test',
+  rootDir: '../../../..',
+  roots: ['<rootDir>/packages/core/notifications/core-notifications-browser'],
+};
