+++ packages/core/notifications/core-notifications-browser/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "@kbn/core-notifications-browser",
+  "private": true,
+  "version": "1.0.0",
+  "main": "./target_node/index.js",
+  "browser": "./target_web/index.js",
+  "license": "SSPL-1.0 OR Elastic License 2.0"
+}
