+++ packages/core/notifications/core-notifications-browser/src/index.ts
@@ -5,15 +5,13 @@
  * in compliance with, at your election, the Elastic License 2.0 or the Server
  * Side Public License, v 1.
  */
-export { NotificationsService } from './notifications_service';
+
 export type {
   ErrorToastOptions,
   ToastOptions,
   Toast,
   ToastInput,
   IToasts,
-  ToastsApi,
   ToastInputFields,
-  ToastsSetup,
-  ToastsStart,
-} from './toasts';
+} from './types';
+export type { ToastsSetup, ToastsStart, NotificationsSetup, NotificationsStart } from './contracts';
